modified:   ../2dplane/2dplane.cpp 	modified:   ../2dplane/2dplane.h 	modified:   ../2dplane/GridStateSpace.cpp 	modified:   ../2dplane/PlaneStateSpace.cpp 	modified:   ../2dplane/obstaclegrid.cpp 	modified:   ../3dplane/3dplane.cpp 	modified:   ../3dplane/3dplane.h 	modified:   ../3dplane/ObstacleGrid3d.cpp 	modified:   ../3dplane/ObstacleGrid3d.h 	modified:   ../README.md 	modified:   ../birrt.h 	modified:   ../birrtstar.h 	deleted:    ../gtest/test_main.cpp 	modified:   ../my_BRRTstar.pro 	modified:   ../my_BRRTstar.pro.user 	modified:   MainWindow.cpp 	modified:   MainWindow.hpp 	modified:   RRTWidget.cpp 	modified:   RRTWidget.hpp 	modified:   main.cpp 	modified:   ../rrtstartree.h

+++ 3dplane/ObstacleGrid3d.h
@@ -24,8 +24,8 @@ public:
      */
     float nearestObstacleDist(const Eigen::Vector3f &state, float maxDist) const;
     void clear();
-    bool &obstacleAt(int x, int y, int z);
-    bool obstacleAt(int x, int y, int z) const;
+    bool &obstacleAt(double x, double y, double z);
+    bool obstacleAt(double x, double y, double z) const;
     bool &obstacleAt(const Eigen::Vector3i &gridLoc);
     bool obstacleAt(const Eigen::Vector3i &gridLoc) const;
